+++ src-docs/src/components/guide_page/guide_page.js
@@ -2,18 +2,37 @@ import PropTypes from 'prop-types';
 import React, { Fragment } from 'react';
 
 import {
+  Link,
+} from 'react-router';
+
+import {
   EuiTitle,
   EuiSpacer,
+  EuiFlexGroup,
+  EuiFlexItem,
+  EuiButton
 } from '../../../../src/components';
 
-export const GuidePage = ({ children, title, intro }) => {
+export const GuidePage = ({ children, title, intro, componentLinkTo }) => {
   return (
     <Fragment>
       <div className="guideSection__text">
-        <EuiTitle size="l">
-          <h1>{title}</h1>
-        </EuiTitle>
-
+        <EuiFlexGroup>
+          <EuiFlexItem>
+            <EuiTitle size="l">
+              <h1>{title}</h1>
+            </EuiTitle>
+          </EuiFlexItem>
+          { componentLinkTo &&
+            <EuiFlexItem grow={false}>
+              <Link to={componentLinkTo}>
+                <EuiButton>
+                  View component code
+                </EuiButton>
+              </Link>
+            </EuiFlexItem>
+          }
+        </EuiFlexGroup>
         <EuiSpacer />
         {intro}
       </div>
@@ -27,4 +46,5 @@ GuidePage.propTypes = {
   children: PropTypes.node,
   title: PropTypes.string,
   intro: PropTypes.node,
+  componentLinkTo: PropTypes.string,
 };