+++ src-docs/src/views/text/text.js
@@ -2,11 +2,12 @@ import React from 'react';
 
 import {
   EuiText,
+  EuiHorizontalRule,
 } from '../../../../src/components';
 
 export default () => (
   <div>
-    <EuiText>
+    <EuiText grow={false}>
       <h1>This is Heading One</h1>
       <p>
         Far out in the uncharted backwaters of the unfashionable end of
@@ -100,6 +101,29 @@ export default () => (
       </p>
 
       <h6>This is Heading Six</h6>
+
+      <EuiHorizontalRule />
+
+      <dl>
+        <dt>
+          The Elder Scrolls: Morrowind
+        </dt>
+        <dd>
+          The opening music alone evokes such strong memories.
+        </dd>
+        <dt>
+          TIE Fighter
+        </dt>
+        <dd>
+          The sequel to XWING, join the dark side and fly for the Emporer.
+        </dd>
+        <dt>
+          Quake 2
+        </dt>
+        <dd>
+          The game that made me drop out of college.
+        </dd>
+      </dl>
     </EuiText>
   </div>
 );